Remove CustomText component
- I've deleted duplicate component
- I've extracted styles in other files
