--- GamerpotFrontend/src/components/CustomText/styles.ts
@@ -1,14 +1,0 @@
-import {StyleSheet} from 'react-native';
-import {colors} from '../../library/constants/colors';
-import {fonts} from '../../library/constants/fonts';
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  text: {
-    color: colors.green,
-  },
-});
