+++ GamerpotFrontend/src/components/DateText/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Text, View} from 'react-native';
 import {styles} from '../styles';
-import {TextWrapper} from '../CustomText/props';
+import {TextWrapper} from '../Tag/props';
 
 const DateText = ({title}: TextWrapper) => {
   return (
