+++ GamerpotFrontend/src/components/Tag/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import {StyleSheet, Text, View} from 'react-native';
-import {colors} from '../../library/constants/colors';
+import {Text, View} from 'react-native';
+import {styles} from './styles';
 
 interface TagProps {
   text: string;
@@ -15,20 +15,3 @@ const Tag = ({text}: TagProps) => {
 };
 
 export default Tag;
-
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: colors.darkPastelGreen,
-    marginTop: 2,
-    borderRadius: 20,
-    paddingHorizontal: 25,
-    paddingVertical: 5,
-    marginRight: 8,
-  },
-  labelTag: {
-    color: colors.secondary,
-    fontSize: 12,
-    fontWeight: '500',
-    lineHeight: 15,
-  },
-});
