+++ GamerpotFrontend/src/components/Tag/props.ts
